Вынес соритьвку в отдкльный модуль
